Removing the sniffer for the database change
This problem resolves Issue #8 .

The sniffer would try to see if the `shipping_method` column of `TABLE_ORDERS` was set to hold 255 characters at least as a `VARCHAR`. Things get complicated if either by another module or some other modification to be something like `TEXT` or `BLOB`.

This change is inconsequential to ZenCart 2.x installations as that field is already set to be a `VARCHAR(255)` by the SQL installer. This modification pretty much ensures the field is at least 255 characters long as the descriptions of the shipping module tends to be pretty long.