Patches out the niche case rate failure
There was a SMALL chance that the quote didn't return. In short, the conditional check was in the wrong place.

Moved the conditional checking for a quote to double-check that there was a valid response. If there isn't, kill it, disable the module, and if you're doing debugging, print that message out.

Removed the Heredocs too, they're just problematic.