Merging new main branch (#55)
* Change api endpoints + removing RevokeToken function (#53)

* Changed API Endpoints and extra

Start of changing to new API handler (still need to figure out why tokens can't be revoked)
Fixed extra slashes in JSON output (characters were still being escaped)

* Removing BearerToken request

- Evidently, revocation requests can only be made for refresh_tokens, not access_tokens.
- Changing the bearerToken to be stored on the system's PHP SESSION variable. The module will now check if there is a value there and if so, assign it, and if not, generate it.

* Repairing Handling Days Counter (#54)

Removing the custom function and instead using the innate val_function from the database. (If the version is 1.5.5 or lower, there is no inherent way to check for the 30-day threshold. So I have to be prepared for that.)

Fixes #52