Domestic and international letters (#56)
* New Version + USPS Letter First-Class
* Merging new main branch (#55)
* Change api endpoints + removing RevokeToken function (#53)
* Changed API Endpoints and extra

Start of changing to new API handler (still need to figure out why tokens can't be revoked)
Fixed extra slashes in JSON output (characters were still being escaped in error logs)

* Removing BearerToken request
  - Evidently, revocation requests can only be made for refresh_tokens, not access_tokens.
  - Changing the bearerToken to be stored on the system's PHP SESSION variable. The module will now check if there is a value there and if so, assign it, and if not, generate it.

* Repairing Handling Days Counter (#54)
Removing the custom function and instead using the innate val_function from the database. (If the version is 1.5.5 or lower, there is no inherent way to check for the 30-day threshold. So I have to be prepared for that.)

Fixes #52

* Adding First Class Mail Letters support
- Changed language files to link to support thread on ZC forums + GitHub repository.
- Separated First Class Mail machineability flags
- Modified data table support for First Class Mail (both installers will now modify)
- Added Endpoint URL and other configuration checks to debug messaging.
- Renamed a bunch of internal variables to differentiate between package processing (anything not First Class Mail) and regular letters processing.

* New language for defines
Just updating the files to be a bit more descriptive.

* Update uspsr.php

Fix for #49
Will check if $body is an array. If it isn't assign NULL to the $bearerToken