Half cart weight fix (#75) + Extras
* Changing math for box order calculation

The total cost of the order should be the number of boxes times the quote.

Additionally, added the number of boxes to the quote so it's more clear.

* Documentation change

* Better formatting of the quote number

This is just in case the USPS API starts trying to reject quotes that don't have the proper number format (ie. too many decimals)