Improve bearer token handling in USPSRestful module
Refactored bearer token initialization to check for a valid token in configuration or session before requesting a new one. Also, bearer tokens are now stored in the session after retrieval for the initial retrieval (after the initial retrieval, it goes to the database instead). Refresh token validation now checks if the constant is defined.  (Because of upgrades.) Removed redundant token check logic from the admin check method.